+++ src/main/kotlin/domain/User.kt
@@ -1,5 +1,7 @@
 package domain
 
+import domain.card.Card
+
 class User(
     name: String,
     cards: MutableList<Card>,
