+++ src/main/kotlin/domain/card/CardShape.kt
@@ -1,0 +1,8 @@
+package domain.card
+
+enum class CardShape(val pattern: String) {
+    HEARTS("하트"),
+    DIAMONDS("다이아몬드"),
+    CLUBS("클로버"),
+    SPADES("스페이드"),
+}
